Introduce type checks into OrtValue and Tensor classes.
