+++ onnxruntime/core/framework/data_types.cc
@@ -5,7 +5,7 @@
 #include "core/framework/tensor.h"
 #include "core/framework/TensorSeq.h"
 #include "core/framework/sparse_tensor.h"
-#include "core/framework/utils.h"
+#include "core/framework/data_types_internal.h"
 #include "core/graph/onnx_protobuf.h"
 
 #ifdef MICROSOFT_AUTOML
