+++ onnxruntime/core/framework/ort_value_tensor_slicer.cc
@@ -82,9 +82,9 @@ void OrtValueTensorSlicer<T>::Iterator::MaterializeMLValue() const {
   // Potential future performance enhancement.
   auto sub_tensor = onnxruntime::make_unique<Tensor>(tensor_data_type_, per_iteration_shape_,
                                              const_cast<void*>(tensor_slice_data_raw), *tensor_location_);
-
+  auto ml_tensor = DataTypeImpl::GetType<Tensor>();
   current_ =
-      OrtValue{sub_tensor.release(), DataTypeImpl::GetType<Tensor>(), DataTypeImpl::GetType<Tensor>()->GetDeleteFunc()};
+      OrtValue{sub_tensor.release(), ml_tensor, ml_tensor->GetDeleteFunc()};
 }
 
 template class OrtValueTensorSlicer<OrtValue>;
