+++ onnxruntime/core/framework/tensor.cc
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 #include "core/framework/tensor.h"
-#include "core/framework/utils.h"
 
 #include <utility>
 #include "core/framework/allocatormgr.h"
