+++ onnxruntime/core/providers/cpu/controlflow/loop.cc
@@ -293,8 +293,9 @@ static OrtValue MakeScalarMLValue(AllocatorPtr& allocator, T value, bool is_1d) 
 
   *p_tensor->MutableData<T>() = value;
 
-  return OrtValue{p_tensor.release(), DataTypeImpl::GetType<Tensor>(),
-                  DataTypeImpl::GetType<Tensor>()->GetDeleteFunc()};
+  auto ml_tensor = DataTypeImpl::GetType<Tensor>();
+  return OrtValue{p_tensor.release(), ml_tensor,
+                  ml_tensor->GetDeleteFunc()};
 }
 
 Status LoopImpl::Initialize() {
