+++ onnxruntime/core/providers/cpu/ml/category_mapper.cc
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/framework/utils.h"
 #include "core/providers/cpu/ml/category_mapper.h"
 #include <algorithm>
 #include <gsl/gsl>
