+++ onnxruntime/core/providers/cpu/tensor/transpose.cc
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 #include "core/providers/cpu/tensor/transpose.h"
-#include "core/framework/utils.h"
 
 namespace onnxruntime {
 
@@ -181,7 +180,7 @@ static Status DoUntypedTranspose(const std::vector<size_t>& permutations, const 
   auto rank = input_shape.NumDimensions();
 
   const auto element_size = input.DataType()->Size();
-  const bool is_string_type = utils::IsDataTypeString(input.DataType());
+  const bool is_string_type = input.IsDataTypeString();
 
   std::vector<size_t> stride(rank);
   for (size_t i = 0; i < rank; i++) {
