+++ onnxruntime/core/providers/cuda/tensor/scatter_elements.cc
@@ -3,7 +3,6 @@
 
 #include "scatter_elements.h"
 #include "scatter_elements_impl.h"
-#include "core/framework/utils.h"
 #include "core/providers/cpu/tensor/utils.h"
 #include "core/providers/common.h"
